rename conn to uascnew
