+++ uascnew/client.go
@@ -1,4 +1,4 @@
-package conn
+package uascnew
 
 import (
 	"fmt"
