+++ uascnew/session.go
@@ -1,4 +1,4 @@
-package conn
+package uascnew
 
 import (
 	"crypto/rand"
